--- src/js/parse.js
@@ -1,3 +1,0 @@
-export default function parse(str) {
-  return Function(` return (${str})`)();
-}
